requirements.txt
Django
django-simple-history